requirements.txt
python-telegram-bot==21.6
requests==2.31.0
segno==1.6.1